% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPid.R
\name{addPid}
\alias{addPid}
\title{Add :: (\code{<package>::<function>}) notation to package functions}
\usage{
addPid(r_script, packages = NULL, ignore = NULL, detach = TRUE)
}
\arguments{
\item{\code{r_script}}{A connection to a file (.R or .Rmd) containing package functions.}

\item{\code{packages}}{Optionally specify a character vector of packages to be searched for.}

\item{\code{ignore}}{Optional character vector of package names to ignore.}

\item{\code{detach}}{Logical stating whether packages should be detached after input. Default is TRUE.}
}
\value{
An edited .R file written to the same directory as input file.
}
\description{
Scan a file of R code for the presence of functions from packages installed on machine or a custom list of packages.
}
\details{
Add :: notation to functions so that packages are directly referenced in their calls.
R package development requires not attaching packages and instead loading them directly by \code{package::function}.
This function and package exists to negate manually appending package identifyers to corresponding functions in R code.
}
\examples{
\dontrun{
addPid("/path/to/r/script")
}

}
\author{
Spencer Seale
}
