% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPid.R
\name{addPid}
\alias{addPid}
\title{Add :: (i.e. \code{<package>::<function>}) notation to package functions}
\usage{
addPid(r_script, packages)
}
\arguments{
\item{\code{r_script}}{A connection to a file (.R or .Rmd) containing package functions.}

\item{\code{packages}}{A character vector of packages having functions that should be searched for in r_script.}
}
\value{
An edited .R or .Rmd file written to the same directory as input file.
}
\description{
Scan a file of R code for the presence of functions from of a list of input packages.
}
\details{
Add :: notation to functions so that packages are directly referenced in their calls.
R package development requires not attaching packages and instead loading them directly by \code{package::function}.
This function and package exists to negate manually appending package identifyers to corresponding functions in R code.
}
\examples{
\dontrun{
pkgs <- c("dplyr", "stringr")

addPid("/path/to/r/script", pkgs)
}

}
\author{
Spencer Seale
}
